requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
alembic
psycopg2-binary
asyncpg
openai
azure-cognitiveservices-search-websearch
azure-search-documents
azure-identity
requests
aiohttp
pytest
httpx
python-jose
passlib
bcrypt
email-validator
python-multipart
mysql-connector-python
azure-search-documents>=11.4.0,<12.0.0
azure-identity>=1.7.0,<2.0.0
